# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HireForward — AI-Native Recruitment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /processes/{id}/candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List candidates for a process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: minScore
          in: query
          schema:
            type: integer
        - name: recommendation
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Add a candidate to a process (generates invite token)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCandidateBody"
      responses:
        "201":
          description: Created candidate with invite link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateWithLink"
  /candidates/{candidateId}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get a candidate with full evaluation report
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Candidate with evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateReport"
        "404":
          description: Not found

  # ── Interview (public — no auth, token-based) ─────────────────────────────
...
  /evaluations/{candidateId}:
    get:
      operationId: getEvaluation
      tags: [evaluations]
      summary: Get evaluation for a candidate
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Evaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evaluation"
        "404":
          description: Not found

  # ── Admin (owner) ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Candidate:
      type: object
      properties:
        id:
          type: string
        jobProcessId:
          type: string
        name:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [invited, started, completed, expired]
        overallScore:
          type: integer
          nullable: true
        recommendation:
          type: string
          nullable: true
        inviteToken:
          type: string
        tokenExpiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, jobProcessId, name, email, status, createdAt]
    CandidateWithLink:
      allOf:
        - $ref: "#/components/schemas/Candidate"
        - type: object
          properties:
            inviteLink:
              type: string
          required: [inviteLink]
    CandidateReport:
      allOf:
        - $ref: "#/components/schemas/Candidate"
        - type: object
          properties:
            evaluation:
              $ref: "#/components/schemas/Evaluation"
            transcript:
              type: array
              items:
                type: object
                properties:
                  role:
                    type: string
                  content:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                required: [role, content]
    CreateCandidateBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        tokenExpiresHours:
          type: integer
          default: 72
      required: [name, email]
...
    Evaluation:
      type: object
      properties:
        id:
          type: string
        candidateId:
          type: string
        overallScore:
          type: integer
        recommendation:
          type: string
        criteriaScores:
          type: array
          items:
            $ref: "#/components/schemas/CriterionScore"
        highlights:
          type: array
          items:
            type: string
        redFlags:
          type: array
          items:
            type: string
        summary:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, candidateId, overallScore, recommendation, criteriaScores, highlights, redFlags, summary, createdAt]
    CriterionScore:
      type: object
      properties:
        name:
          type: string
        score:
          type: integer
        max:
          type: integer
        weight:
          type: integer
        justification:
          type: string
      required: [name, score, max, weight, justification]
...