# lib/api-zod/src/generated/types/createCandidateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireForward — AI-Native Recruitment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCandidateBody {
  name: string;
  email: string;
  tokenExpiresHours?: number;
}